move utils::list_s3_paths to new s3 module, update references in other files
