+++ api/scpca_portal/s3.py
@@ -1,4 +1,5 @@
 import subprocess
+from collections import namedtuple
 from pathlib import Path
 
 from django.conf import settings
@@ -101,3 +102,72 @@ def create_download_url(self):
             },
             ExpiresIn=60 * 60 * 24 * 7,  # 7 days in seconds.
         )
+
+
+"""
+The `aws s3 ls <bucket>` command called in `list_s3_paths()` returns a list of two types of entries:
+- Bucket Object Entries
+- Bucket Prefix Entries
+In order to create a standard API, where `entry.file_path` could be accessed
+irrespective of the entry type, we've created two named tuples which follow the return format
+of each of the bucket entry types.
+"""
+BucketObjectEntry = namedtuple("BucketObjectEntry", ["date", "time", "size_in_bytes", "file_path"])
+BucketPrefixEntry = namedtuple("BucketPrefixEntry", ["prefix_designation", "file_path"])
+
+
+def list_s3_paths(
+    relative_path: Path = Path(),
+    *,
+    bucket_path: Path = Path(common.INPUT_BUCKET_NAME),
+    recursive: bool = True,
+):
+    """
+    Queries a path on an inputted s3 bucket
+    and returns bucket's existing content as a list of Path objects,
+    relative to (without) the bucket prefix.
+
+    The `aws s3 ls <bucket>` command returns a list of two types of entries:
+    - Bucket Object Entries
+    - Bucket Prefix Entries
+    In order to create a standard API, where `entry.file_path` could be accessed
+    irrespective of the entry type, we've created two named tuples which follow the return format
+    of each of the bucket entry types.
+    """
+    root_path = Path(*bucket_path.parts, *relative_path.parts)
+    command_inputs = ["aws", "s3", "ls", f"s3://{root_path}"]
+
+    if recursive:
+        command_inputs.append("--recursive")
+
+    if "public" in str(bucket_path):
+        command_inputs.append("--no-sign-request")
+
+    try:
+        result = subprocess.run(command_inputs, capture_output=True, text=True, check=True)
+        output = result.stdout
+    except subprocess.CalledProcessError as error:
+        logger.error(
+            """
+            `{}`: Cause of error not returned, note: folder must exist and be non-empty
+            """.format(
+                error
+            )
+        )
+        return []
+
+    bucket_entries = []
+    for line in output.splitlines():
+        if line.strip().startswith("PRE"):
+            bucket_entries.append(BucketPrefixEntry._make(line.split()))
+        else:
+            bucket_entries.append(BucketObjectEntry._make(line.split()))
+
+    file_paths = [Path(entry.file_path) for entry in bucket_entries]
+
+    # recursive returns an absolute path (see docstring)
+    if recursive:
+        bucket_keys = Path(*bucket_path.parts[1:])
+        return [entry.relative_to(bucket_keys) for entry in file_paths]
+
+    return file_paths
